+++ src/simulator/simulator.h
@@ -31,12 +31,12 @@
 	 */
 	struct thread
 	{
-		unsigned tid;      /**< Thread ID.                  */
-		unsigned workload; /**< Total assigned workload.    */
-		unsigned ntasks;   /**< Number of assigned tasks.   */
-		double avg;        /**< Average task size.          */
-		unsigned max;      /**< Maximum assigned task size. */
-		unsigned min;      /**< Minimum assigned task size. */
+		unsigned tid;    /**< Thread ID.                  */
+		double workload; /**< Total assigned workload.    */
+		unsigned ntasks; /**< Number of assigned tasks.   */
+		double avg;      /**< Average task size.          */
+		double max;      /**< Maximum assigned task size. */
+		double min;      /**< Minimum assigned task size. */
 	};
 
 	/**
@@ -52,7 +52,7 @@
 	/**
 	 * @brief Initializes the scheduling strategy.
 	 */
-	typedef void (*scheduler_init_t)(const unsigned *, unsigned, unsigned);
+	typedef void (*scheduler_init_t)(const double *, unsigned, unsigned);
 	
 	/**
 	 * @brief Scheduling strategy.
@@ -66,11 +66,11 @@
 	
 	/* Forward definitions. */
 	extern struct thread *threads;
-	extern void schedule(const unsigned *, unsigned, unsigned, unsigned);
-	extern void scheduler_static_init(const unsigned *, unsigned, unsigned);
+	extern void schedule(const double *, unsigned, unsigned, unsigned);
+	extern void scheduler_static_init(const double *, unsigned, unsigned);
 	extern unsigned scheduler_static_sched(unsigned);
 	extern void scheduler_static_end(void);
-	extern void scheduler_dynamic_init(const unsigned *, unsigned, unsigned);
+	extern void scheduler_dynamic_init(const double *, unsigned, unsigned);
 	extern unsigned scheduler_dynamic_sched(unsigned);
 	extern void scheduler_dynamic_end(void);
 
@@ -91,7 +91,7 @@
 	struct dqueue
 	{
 		unsigned tid;        /**< Thread ID.                */
-		unsigned remaining;  /**< Remaining time.           */
+		double remaining;    /**< Remaining time.           */
 		struct dqueue *next; /**< Next thread in the queue. */
 	};
